==> models/NER/reproduce/00-gather_mangas_data.ipynb <==
Собирает в единую табличку ссылки на файлы сырых скачанных данных

from pathlib import Path
from collections import defaultdict
import pandas as pd
import json
import os
from tqdm.notebook import tqdm
import matplotlib.pyplot as plt

root = Path('../../../data/NER/')
data_root = root / 'raw'
processed_data_root = root / 'processed'

# Add data from grouple manga main page

# data example
data = json.load(open(data_root / 'grouple_manga_pages/0.json'))
data

# data example
data = json.load(open(data_root / 'grouple_manga_volumes/0.json'))
# data

# Create url to page_file and volume_file

url2page = dict()
path = data_root / 'grouple_manga_pages/'

for file in tqdm(os.listdir(path)):
    data = json.load(open(path / file))
    url_title = data['url'].split('/')[-1]
    url2page[url_title] = file

# get several volumes per one manga
url2volumes = defaultdict(lambda: [])
path = data_root / 'grouple_manga_volumes/'

for file in tqdm(os.listdir(path)):
    if not file.endswith('.json'):
        continue
    data = json.load(open(path / file))
    url_title = data['url'].split('/')[-3]
    url2volumes[url_title].append(file)


# merge volumes into one file
url2volume = defaultdict(lambda: None)
merged_files_path = data_root / 'grouple_merged_volumes/'

for i, (url_title, vols) in tqdm(enumerate(url2volumes.items()), total=len(url2volumes)):
    obj = {'comments': [],
           'links': []}
    for vol in vols:
        data = json.load(open(path / vol))
        obj['comments'] += data['comments']
        obj['links'] += data['links']
    obj['url'] = data['url']
    
    filename = f'{i}.json'
    url2volume[url_title] = filename
    json.dump(obj, open(merged_files_path / filename, 'w'))

# Create df with links to files

records = []
for url in tqdm(list(url2page.keys())):
    page = url2page[url]
    volumes = url2volume[url]
    records.append((url, page, volumes))

df = pd.DataFrame.from_records(records, columns=['url', 'page_file', 'volumes_data_file'])

df.to_csv(processed_data_root / 'merged_grouple_data.csv', index=False)

# Plotsurl2volume

# How many volumes loaded per manga
counts = list(map(len, url2volumes.values()))

counts = pd.Series(list(manga2volumes.values()))
_ = plt.hist(counts, bins=100)
plt.showz()

_ = plt.hist(counts[counts < 50], bins=50)
plt.show()
