==> models/NER/reproduce/01-manual_data_markup.ipynb <==
from pathlib import Path
import pandas as pd
import json
import matplotlib.pyplot as plt
from tqdm.notebook import tqdm 

from pairing_utils import get_pairing
from markup_data import Manga

root = Path('../../../data/NER/')
raw_root = root / 'raw'
processed_root = root / 'processed'

df = pd.read_csv(processed_root / 'merged_grouple_data.csv')
print('all', df.shape)
print('with comments', df[~df.volumes_data_file.isna()].shape)
df = df[~df.volumes_data_file.isna()]
df.head()

# Load comments for each manga

def load_data(page_file, volumes_file):
    page = raw_root / 'grouple_manga_pages/' / page_file
    volumes = raw_root / 'grouple_merged_volumes/' / volumes_file
    return {'page': json.load(open(page)),
            'volumes': json.load(open(volumes))}

dataset = []
N = 100

# def hotfix(pname):
#     for i, c in enumerate(pname[1:], start=1):
#         if c.isupper():
#             return pname[:i]
#     return pname
        
for name, page, volumes in tqdm(df.to_records(index=False)[:N]):
    data = load_data(page, volumes)
    url = data['page']['url']
    descr = data['page']['description']
    comments = data['volumes']['comments']
    manga = Manga(comments)
    dataset.append(manga)

comms = dataset[0].comments.comments
comms[:10]

import random
from typing import List

class Saver:
    """Class for sampling and saving labled comments data"""
    def __init__(self, save_path: str, dataset: list):
        self.path = save_path
        self.dataset = dataset
        self.used = set()
    
    def sample(self) -> List[str, ]:
        # Fetches sample from data
        def get_index():
            m = random.randint(0, len(self.dataset)-1)
            c = random.randint(0, len(self.dataset[m].comments.comments)-1)
            return m, c
        
        m, c = get_index()
        idx = f'{m}_{c}'
        while idx in self.used:
            m, c = get_index()
            idx = f'{m}_{c}'
            
        self.used.add(idx)
        comm = self.dataset[m].comments.comments[c]
        return comm.split()
        
    def add(self, words: List[str, ], labels: List[str, ]) -> None:
        # Writes out labled sample
        with open(self.path, 'a') as f:
            for w, l in zip(words, labels):
                f.write(f'{w} _ _ {l}\n')
            f.write('\n')
            
saver = Saver(processed_root / 'comments/raw.txt', dataset)

# Manually label i-th word(s) as PER(son)

while 1:
    words = saver.sample()
    labels = ['O']*len(words)
    s = [(i,w) for i,w in enumerate(words)]
    s = '\n'.join(list(map(lambda p: '%d %s' % p, s)))
    print(s)
    ixs = input()
    if len(ixs) > 0:
        ixs = list(map(int, ixs.split()))
        for ix in ixs:
            labels[ix] = 'PER'
    saver.add(words, labels)
    print('-----------------')

len(open(saver.path).read().split('\n\n'))
