==> models/NER/reproduce/02-augment_comments.ipynb <==
from pathlib import Path
from copy import deepcopy
from random import choice, shuffle
import pandas as pd
import json
import matplotlib.pyplot as plt

root = Path('../../../data/NER/')
raw_root = root / 'raw'
processed_root = root / 'processed'

ydf1 = pd.read_csv(raw_root / 'you_anime_characters_refs.csv', sep='\t')
ydf1['anime_url'] = ydf1.page_link.apply(lambda s: s.split('/')[-1])

ydf2 = pd.read_csv(raw_root / 'you_anime_refs.csv', sep='\t')
ydf2['anime_url'] = ydf2.link.apply(lambda s: s.split('/')[-1])

ydf = pd.merge(ydf1, ydf2, on='anime_url')
ydf = ydf['title name'.split()]

ydf.title = ydf.title.apply(str.strip)
ydf.name = ydf.name.apply(str.strip)

ydf.head(3)

names = ydf.name.drop_duplicates().tolist()
print(len(names))
names[:5]

# Read data

def read_conll(path):
    """
    Returns list of sentences. Sentence is a list of pairs (word, NE-label)
    """
    data = open(path).read().split('\n\n')[1:]
    data = list(map(str.splitlines, data))
    get_token_and_ne = lambda s: (s.split()[0], s.split()[-1])
    data = list(map(lambda sent: list(map(get_token_and_ne, sent)), data))
    return data

sents = read_conll(processed_root / 'comments/original/raw.txt')
with_character = lambda sent: any(map(lambda token: token[-1] == 'PER', sent))

sents_characters = list(filter(with_character, sents))
len(sents_characters) / len(sents)

# Augment

def get_rand_character(characters):
    char = choice(characters)
    if len(char.split()) > 1:
        char = choice(char.split())
    return char

augmented_sents_characters = []
AUGMENTATION_ROUNDS = 10

for i in range(AUGMENTATION_ROUNDS):
    for sent in sents_characters:
        sent = deepcopy(sent)
        # augment sentence
        for i, token in enumerate(sent):
            if token[-1] == 'PER':
                token = list(token)
                token[0] = get_rand_character(names)
                sent[i] = tuple(token)
        augmented_sents_characters.append(sent)

len(augmented_sents_characters)

# Mix data with augmented and save

mixed = augmented_sents_characters + sents
shuffle(mixed)

token_to_str = lambda token: '%s _ _ %s' % token
sent_to_str_tokens = lambda sent: '\n'.join(list(map(token_to_str, sent)))

mixed = list(map(sent_to_str_tokens, mixed))
with open(processed_root / f'comments/augmented_{AUGMENTATION_ROUNDS}/raw.txt', 'w') as f:
    f.write('\n\n'.join(mixed))
